Temperature fetching now functinal.
I've got a few basic commands that run to fetch the temperature, as
well as all the location data for each of my sibs.

I've also started work on the "dial" class. This will abstract away
the process of working with the dial face. I'll be able to load the
dial face. tell it "Go to this temperature" then it will move and
save it's new location.