I've got 1 dial working!
That's about 90% of the work for one other, and 60% of the work for
four others. I've got to master communication between the arduino
and the node before I can get all the dials running, but reading
and writing the dial information is working, moving the dial is
working. Progress! As slow as it maybe.