Hardware is all worked out.
Now it is just a matter of getting all the software sorted. I've
come a long way with that this evening, but there is still plenty
to do. A number of the needles seem to be wired backwards for
example. I also need to get it wired properly into wunderground so
it's pulling in actual data.

But it's almost done! HUZZAH! Almost a year late, and it probably
won't be able to run on battery, but it's almost done!