Software is done!
I believe I have completed the software for this project. They
dials all move, in the correct direction, and in both directions.
You can update them by connecting to the web interface of the
device. You can also see what the device thinks it's currently
pointing at.

I set the number of refreshes down to 48 a day. This means it makes
a total of 288 requests to wunderground, well under the 500 query
limit.

What's left is to try and nice up the cords a bit and make a new
face for it. Some of the dials get completely cropped off of the
device's face. So it is necessary to shink, crop, or otherwise
alter the dial faces. Next time I work on this.

I hope to some day revisit this project and give mom and upgrade.
but before that I'll need to learn how to make custom PCBs. I
hopefully then will also be able to master sleep mode so that the
device can be battery operated.